AndriiN. To prepare for this chapter, I am going to create a new project named Platform, using the template that provides the minimal ASP.NET Core setup (ASP.NET Core empty).
Open the Platform.sln file in the Platform folder. Select
Project ➤ Platform Properties, navigate to the Debug page, and change
the App URL field to http://localhost:5000, as shown in Figure 12-1.
This changes the port that will be used to receive HTTP requests.
Select File ➤ Save All to save the configuration changes.