AndriiN This command adds the Swashbuckle.AspNetCore package to the project. You will see this package used in Chapter 20, but for now, it is the effect of the dotnet add package command that is important.
If you are using Visual Studio, you can add the package by right-clicking
the Platform item in the Solution Explorer and selecting Manage NuGet
packages from the popup menu. Click Browse and enter Swashbuckle.AspNetCore
into the search text box. Select the Swashbuckle.AspNetCore package from
the list, choose the 6.4.0 version from the Version drop-down list,
and click the Install button, as shown in Figure 12-7. You will be
prompted to accept the license of the package and its dependencies.